combined jsonb tests and deleted duplicated transient files
