+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/core/JDTServicesManager.java
@@ -33,8 +33,7 @@ import org.eclipse.lsp4jakarta.jdt.core.annotations.AnnotationDiagnosticsCollect
 import org.eclipse.lsp4jakarta.jdt.core.beanvalidation.BeanValidationDiagnosticsCollector;
 import org.eclipse.lsp4jakarta.jdt.core.jax_rs.ResourceMethodDiagnosticsCollector;
 import org.eclipse.lsp4jakarta.jdt.core.jax_rs.Jax_RSClassDiagnosticsCollector;
-import org.eclipse.lsp4jakarta.jdt.core.jsonb.JsonbCreatorDiagnosticsCollector;
-import org.eclipse.lsp4jakarta.jdt.core.jsonb.JsonbTransientDiagnosticsCollector;
+import org.eclipse.lsp4jakarta.jdt.core.jsonb.JsonbDiagnosticsCollector;
 import org.eclipse.lsp4jakarta.jdt.core.jsonp.JsonpDiagnosticCollector;
 import org.eclipse.lsp4jakarta.jdt.core.cdi.ManagedBeanDiagnosticsCollector;
 import org.eclipse.lsp4jakarta.jdt.core.di.DependencyInjectionDiagnosticsCollector;
@@ -78,8 +77,7 @@ public class JDTServicesManager {
         diagnosticsCollectors.add(new PersistenceMapKeyDiagnosticsCollector());
         diagnosticsCollectors.add(new ResourceMethodDiagnosticsCollector());
         diagnosticsCollectors.add(new Jax_RSClassDiagnosticsCollector());
-        diagnosticsCollectors.add(new JsonbCreatorDiagnosticsCollector());
-        //diagnosticsCollectors.add(new JsonbTransientDiagnosticsCollector());
+        diagnosticsCollectors.add(new JsonbDiagnosticsCollector());
         diagnosticsCollectors.add(new ManagedBeanDiagnosticsCollector());
         diagnosticsCollectors.add(new DependencyInjectionDiagnosticsCollector());
         diagnosticsCollectors.add(new JsonpDiagnosticCollector());
