+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/src/main/java/org/eclipse/lsp4jakarta/jdt/jsonb/JsonbDiagnosticsCollectorTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright (c) 2022 IBM Corporation and others.
+* Copyright (c) 2021 IBM Corporation and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Eclipse Public License v. 2.0 which is available at
@@ -8,7 +8,7 @@
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
-*     Adit Rada - initial API and implementation
+*     IBM Corporation - initial API and implementation
 *******************************************************************************/
 package org.eclipse.lsp4jakarta.jdt.jsonb;
 
@@ -34,9 +34,44 @@ import org.eclipse.lsp4jakarta.jdt.core.BaseJakartaTest;
 import org.eclipse.lsp4jakarta.jdt.core.JDTUtils;
 import org.junit.Test;
 
-public class JsonbTransientFieldDiagnosticTest extends BaseJakartaTest {
+public class JsonbDiagnosticsCollectorTest extends BaseJakartaTest {
     protected static JDTUtils JDT_UTILS = new JDTUtils();
 
+    @Test
+    public void deleteExtraJsonbCreatorAnnotation() throws Exception {
+        JDTUtils utils = JDT_UTILS;
+        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
+        IFile javaFile = javaProject.getProject().getFile(
+                new Path("src/main/java/io/openliberty/sample/jakarta/jsonb/ExtraJsonbCreatorAnnotations.java"));
+        String uri = javaFile.getLocation().toFile().toURI().toString();
+
+        JakartaDiagnosticsParams diagnosticsParams = new JakartaDiagnosticsParams();
+        diagnosticsParams.setUris(Arrays.asList(uri));
+
+        Diagnostic d1 = d(18, 11, 39,
+                "Only one constructor or static factory method can be annotated with @JsonbCreator in a given class.",
+                DiagnosticSeverity.Error, "jakarta-jsonb", "MultipleJsonbCreatorAnnotations");
+        
+        Diagnostic d2 = d(21, 48, 61,
+                "Only one constructor or static factory method can be annotated with @JsonbCreator in a given class.",
+                DiagnosticSeverity.Error, "jakarta-jsonb", "MultipleJsonbCreatorAnnotations");
+
+        assertJavaDiagnostics(diagnosticsParams, utils, d1, d2);
+
+        // test code actions
+        JakartaJavaCodeActionParams codeActionParams1 = createCodeActionParams(uri, d1);
+        TextEdit te1 = te(17, 4, 18, 4, "");
+        CodeAction ca1 = ca(uri, "Remove @JsonbCreator", d1, te1);
+        
+        assertJavaCodeAction(codeActionParams1, utils, ca1);
+
+        JakartaJavaCodeActionParams codeActionParams2 = createCodeActionParams(uri, d2);
+        TextEdit te2 = te(20, 4, 21, 4, "");
+        CodeAction ca2 = ca(uri, "Remove @JsonbCreator", d2, te2);
+
+        assertJavaCodeAction(codeActionParams2, utils, ca2);
+    }
+    
     @Test
     public void diagnosticWhenJsonbTransientIsNotMutuallyExclusive() throws Exception {
         JDTUtils utils = JDT_UTILS;
