--- jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/src/main/java/org/eclipse/lsp4jakarta/jdt/jsonb/JsonbTest.java
@@ -1,74 +1,0 @@
-/*******************************************************************************
-* Copyright (c) 2021 IBM Corporation and others.
-*
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v. 2.0 which is available at
-* http://www.eclipse.org/legal/epl-2.0.
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Contributors:
-*     IBM Corporation - initial API and implementation
-*******************************************************************************/
-package org.eclipse.lsp4jakarta.jdt.jsonb;
-
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.assertJavaCodeAction;
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.assertJavaDiagnostics;
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.ca;
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.createCodeActionParams;
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.d;
-import static org.eclipse.lsp4jakarta.jdt.core.JakartaForJavaAssert.te;
-
-import java.util.Arrays;
-
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.runtime.Path;
-import org.eclipse.jdt.core.IJavaProject;
-import org.eclipse.lsp4j.CodeAction;
-import org.eclipse.lsp4j.Diagnostic;
-import org.eclipse.lsp4j.DiagnosticSeverity;
-import org.eclipse.lsp4j.TextEdit;
-import org.eclipse.lsp4jakarta.commons.JakartaDiagnosticsParams;
-import org.eclipse.lsp4jakarta.commons.JakartaJavaCodeActionParams;
-import org.eclipse.lsp4jakarta.jdt.core.BaseJakartaTest;
-import org.eclipse.lsp4jakarta.jdt.core.JDTUtils;
-import org.junit.Test;
-
-public class JsonbTest extends BaseJakartaTest {
-    protected static JDTUtils JDT_UTILS = new JDTUtils();
-
-    @Test
-    public void deleteExtraJsonbCreatorAnnotation() throws Exception {
-        JDTUtils utils = JDT_UTILS;
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                new Path("src/main/java/io/openliberty/sample/jakarta/jsonb/ExtraJsonbCreatorAnnotations.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaDiagnosticsParams diagnosticsParams = new JakartaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        Diagnostic d1 = d(18, 11, 39,
-                "Only one constructor or static factory method can be annotated with @JsonbCreator in a given class.",
-                DiagnosticSeverity.Error, "jakarta-jsonb", "MultipleJsonbCreatorAnnotations");
-        
-        Diagnostic d2 = d(21, 48, 61,
-                "Only one constructor or static factory method can be annotated with @JsonbCreator in a given class.",
-                DiagnosticSeverity.Error, "jakarta-jsonb", "MultipleJsonbCreatorAnnotations");
-
-        assertJavaDiagnostics(diagnosticsParams, utils, d1, d2);
-
-        // test code actions
-        JakartaJavaCodeActionParams codeActionParams1 = createCodeActionParams(uri, d1);
-        TextEdit te1 = te(17, 4, 18, 4, "");
-        CodeAction ca1 = ca(uri, "Remove @JsonbCreator", d1, te1);
-        
-        assertJavaCodeAction(codeActionParams1, utils, ca1);
-
-        JakartaJavaCodeActionParams codeActionParams2 = createCodeActionParams(uri, d2);
-        TextEdit te2 = te(20, 4, 21, 4, "");
-        CodeAction ca2 = ca(uri, "Remove @JsonbCreator", d2, te2);
-
-        assertJavaCodeAction(codeActionParams2, utils, ca2);
-    }
-}
